Pixel grid writes to NeXus incorrectly
